+++ AMAPVox/AMAPVoxCore/src/main/java/fr/amap/lidar/amapvox/voxelisation/configuration/params/EchoesWeightByRankParams.java
@@ -15,27 +15,27 @@ public class EchoesWeightByRankParams {
      * Weighting matrix as a 2d array, first number of shot's echoes then echo
      * rank
      */
-    private final float[][] weightingData;
+    private final double[][] weightingData;
 
-    public final static float[][] DEFAULT_ALS_WEIGHTING = new float[][]{
-        {1.00f, Float.NaN, Float.NaN, Float.NaN, Float.NaN, Float.NaN, Float.NaN},
-        {0.62f, 0.38f, Float.NaN, Float.NaN, Float.NaN, Float.NaN, Float.NaN},
-        {0.40f, 0.35f, 0.25f, Float.NaN, Float.NaN, Float.NaN, Float.NaN},
-        {0.28f, 0.29f, 0.24f, 0.19f, Float.NaN, Float.NaN, Float.NaN},
-        {0.21f, 0.24f, 0.21f, 0.19f, 0.15f, Float.NaN, Float.NaN},
-        {0.16f, 0.21f, 0.19f, 0.18f, 0.14f, 0.12f, Float.NaN},
-        {0.15f, 0.17f, 0.15f, 0.16f, 0.12f, 0.19f, 0.06f}};
+    public final static double[][] DEFAULT_ALS_WEIGHTING = new double[][]{
+        {1.00d, Double.NaN, Double.NaN, Double.NaN, Double.NaN, Double.NaN, Double.NaN},
+        {0.62d, 0.38d, Double.NaN, Double.NaN, Double.NaN, Double.NaN, Double.NaN},
+        {0.40d, 0.35d, 0.25d, Double.NaN, Double.NaN, Double.NaN, Double.NaN},
+        {0.28d, 0.29d, 0.24d, 0.19d, Double.NaN, Double.NaN, Double.NaN},
+        {0.21d, 0.24d, 0.21d, 0.19d, 0.15d, Double.NaN, Double.NaN},
+        {0.16d, 0.21d, 0.19d, 0.18d, 0.14d, 0.12d, Double.NaN},
+        {0.15d, 0.17d, 0.15d, 0.16d, 0.12d, 0.19d, 0.06f}};
 
-    public final static float[][] DEFAULT_TLS_WEIGHTING = new float[][]{
-        {1.00f, Float.NaN, Float.NaN, Float.NaN, Float.NaN, Float.NaN, Float.NaN},
-        {0.50f, 0.50f, Float.NaN, Float.NaN, Float.NaN, Float.NaN, Float.NaN},
-        {1 / 3.0f, 1 / 3.0f, 1 / 3.0f, Float.NaN, Float.NaN, Float.NaN, Float.NaN},
-        {0.25f, 0.25f, 0.25f, 0.25f, Float.NaN, Float.NaN, Float.NaN},
-        {0.20f, 0.20f, 0.20f, 0.20f, 0.20f, Float.NaN, Float.NaN},
-        {1 / 6.0f, 1 / 6.0f, 1 / 6.0f, 1 / 6.0f, 1 / 6.0f, 1 / 6.0f, Float.NaN},
-        {1 / 7.0f, 1 / 7.0f, 1 / 7.0f, 1 / 7.0f, 1 / 7.0f, 1 / 7.0f, 1 / 7.0f}};
+    public final static double[][] DEFAULT_TLS_WEIGHTING = new double[][]{
+        {1.00d, Double.NaN, Double.NaN, Double.NaN, Double.NaN, Double.NaN, Double.NaN},
+        {0.50d, 0.50d, Double.NaN, Double.NaN, Double.NaN, Double.NaN, Double.NaN},
+        {1 / 3.0d, 1 / 3.0d, 1 / 3.0d, Double.NaN, Double.NaN, Double.NaN, Double.NaN},
+        {0.25d, 0.25d, 0.25d, 0.25d, Double.NaN, Double.NaN, Double.NaN},
+        {0.20d, 0.20d, 0.20d, 0.20d, 0.20d, Double.NaN, Double.NaN},
+        {1 / 6.0d, 1 / 6.0d, 1 / 6.0d, 1 / 6.0d, 1 / 6.0d, 1 / 6.0d, Double.NaN},
+        {1 / 7.0d, 1 / 7.0d, 1 / 7.0d, 1 / 7.0d, 1 / 7.0d, 1 / 7.0d, 1 / 7.0f}};
 
-    public EchoesWeightByRankParams(float[][] weightingData) {
+    public EchoesWeightByRankParams(double[][] weightingData) {
         this.weightingData = weightingData;
     }
 
@@ -44,7 +44,7 @@ public class EchoesWeightByRankParams {
      * @return Weighting matrix as a 2d array, first by number of shot's echoes
      * then echo rank
      */
-    public float[][] getWeightingData() {
+    public double[][] getWeightingData() {
         return weightingData;
     }
 }
